2023.08.21 서류/면접/최종 합격결과 조회 및 저장

+++ src/main/java/com/wizian/admission/wizianb/domain/PassManagement.java
@@ -21,7 +21,7 @@ public class PassManagement {
     private String FnlPassYn;
 
     // 전형평가 대상자
-    private int evAvgScore;
+    private float evAvgScore;
     private int prelOrd;
 
 }
